requirements.txt
# Django app
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
waitress==3.0.2
whitenoise==6.9.0
# Orchestrator
azure-mgmt-resource==24.0.0
azure-mgmt-compute==37.0.0
azure-mgmt-network==29.0.0
azure-identity==1.25.0
pytz==2025.2